[Android] CustomTabToolbar moved to customtabs/features/toolbar
Chromium change:
https://chromium.googlesource.com/chromium/src/+/6fd2d9fe48d63008fa3c89d2bb666f1432fc54ee

Toolbar: Move CustomTabToolbar to customtabs/features/toolbar

Class CustomToolbar inherits ToolbarLayout for the layout for
Custom Tab Toolbar. It also provides an implementation of
LocationBar interface used in omnibox. The concrete
implementation classes for Custom Tab don't necessarily have
to be a part of toolbar/. This CL moves the class to
customtab/features/toolbar where other implementation of Custom
Tab features are placed. This also helps reduce the outbound
dependencies the class brought in to toolbar/.

Bug: 1127732