[Android] ThemeColorProvider moved to toolbar/
Chromium change:
https://chromium.googlesource.com/chromium/src/+/73ff9e7a8f7b2a59009fefa3f64dd943199a55a5

Toolbar: Move ThemeColorProvider to toolbar/

ThemeColorProvider deals with colors on toolbar for a
suitable theme. It is mainly used within toolbar/, only by
a few outside the package(namely task.tab_management). This
CL moves the class to toolbar/ which it can be justified to
be a part of.

Bug: 1127732